Create scoring_epochs.json
To simplify scoring calculation, I'd like to maintain this json file to define the timestamps for when to start / stop applying scores for notarisations.
Each timestamp relates to the UNIX epoch time of the relevant commit hash of the merge into master branch when a coin was added or removed from dPoW as noted in an updateXXX.md file. 

For coins being added, the timestamp of the commit is used. Notaries should be aware of an pending change if subscribed to this repo as per https://github.com/KomodoPlatform/dPoW/blob/dev/doc/bible.md#nn-duties section iv. The earliest to update and add new coins will benefit from bonus notarisation points for their rapid response.

For coins being removed, an additional 24 hours is proposed as per section iv. `react in a timely manner (<24h)`  so the timestamp is commit hash epoch time + 24hrs.